Treat rapids-dask-dependency CI artifacts as pure wheels (#30)
## Description
This marks `rapids-dask-dependency` as a pure wheel, meaning that the CI
artifacts are not specific to a Python version or CPU architecture. This
change depends on https://github.com/rapidsai/gha-tools/pull/96, and
makes CI workflows more robust by allowing the test matrix to be
separated from the build matrix.