Fix CLI invocations (#46)
Module loads via `python -m *` take a different code path than direct usage of the dask/distributed CLI. The finder is still run to get the spec, but instead of the loader's `[create|exec]_module` methods being called next the loader's `get_code` method is called to simply load the `__main__` module as a text file to be executed directly. To support this use case, this PR splits up the finder and the loader. The loader is now a SourceFileLoader that is instantiated once per file to be loaded, allowing it to support loading the text of that file on demand by command line invocations. In order for this to work, the spec needs to be properly configured before the finder's `find_spec` call returns. The `update_spec` function I had in place before was a bit of a hack. This PR replaces that with a call to `importlib.util.find_spec` on the real module to get the full spec right up front. This allowed me to do quite a bit of other simplification; since the spec is now correct up front, I removed `update_spec` entirely and converted the loaders from stateful classes that updated the spec on load to simple factories since the spec is already guaranteed to be correct. Finally, I added some new tests of the CLI, module usage, and the accessor module.

Closes #45 
Closes https://github.com/rapidsai/docker/issues/668

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Richard (Rick) Zamora (https://github.com/rjzamora)

URL: https://github.com/rapidsai/rapids-dask-dependency/pull/46